Move util to utils file
