+++ src/new_routes/courses/[courseId]/leaderboard/overall.tsx
@@ -9,22 +9,12 @@ import default_avatar from 'src/assets/default-avatar.jpg';
 import { useTypedSelector } from 'src/commons/utils/Hooks';
 import LeaderboardActions from 'src/features/leaderboard/LeaderboardActions';
 import type { LeaderboardRow } from 'src/features/leaderboard/LeaderboardTypes';
+import { convertToRandomNumber } from 'src/pages/leaderboard/subcomponents/LeaderboardUtils';
 
 import leaderboardBackground from '../../../../../assets/leaderboard_background.jpg';
 import LeaderboardDropdown from '../../../../pages/leaderboard/subcomponents/LeaderboardDropdown';
 import LeaderboardExportButton from '../../../../pages/leaderboard/subcomponents/LeaderboardExportButton';
 import LeaderboardPodium from '../../../../pages/leaderboard/subcomponents/LeaderboardPodium';
-
-// Set sample profile pictures (Seeded random)
-export function convertToRandomNumber(id: string): number {
-  const str = id.slice(1);
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    const char = str.charCodeAt(i);
-    hash = (hash << 5) - hash + char;
-  }
-  return (Math.abs(hash) % 7) + 1;
-}
 
 const columnDefs: ColDef<LeaderboardRow>[] = [
   {
