+++ src/pages/leaderboard/subcomponents/LeaderboardUtils.ts
@@ -44,3 +44,13 @@ export const contestLeaderboardLoader: LoaderFunction = ({ request, params }) =>
   // Pass through and proceed
   return null;
 };
+
+export function convertToRandomNumber(id: string): number {
+  const str = id.slice(1);
+  let hash = 0;
+  for (let i = 0; i < str.length; i++) {
+    const char = str.charCodeAt(i);
+    hash = (hash << 5) - hash + char;
+  }
+  return (Math.abs(hash) % 7) + 1;
+}
